Handle systems with non-existent sensors.
In virtual environments, sensors may not be available (e.g. qemu, "sensors -u" and "sensors-detect" show zero found sensors; but other vms may expose the system sensors of course).

This results in non-digit output (typically empty string), so this patch checks if the sensor output is a non-zero string, and if so, it must start with a digit to try conversion.

Alternatively, we could just have try/except to catch the float() parse error.

Only tested in vm - should also be tested on real node (e.g. change 100 to 10 and check that it correctly goes to maintenance)